Plugin IDs API route
